requirements.txt
torch 
torchvision 
torchaudio
mediapipe
opencv-python
pillow
eva-decord
supervision